reorganize Utilities into Loops, Conditionals, and Time
