+++ libs/core/src/lib/actions/files.ts
@@ -2,7 +2,7 @@ import { BotFileOptions, BotAction, BotFilesAction } from "../interfaces"
 import { enrichBotFileOptionsWithDefaults, getFileUrl } from "../helpers/files"
 import { inject } from "./inject"
 import { errors } from "./errors"
-import { forAll } from "./utilities"
+import { forAll } from "./loops"
 import { goTo } from "./navigation"
 import { unpipeInjects } from "../helpers/pipe"
 import { AbortLineSignal, PipeValue } from "../types"
