+++ libs/core/src/lib/actions/loops.ts
@@ -13,37 +13,6 @@ import { PipeValue } from '../types/pipe-value'
 import { AbortLineSignal, isAbortLineSignal } from '../types/abort-line-signal'
 import { processAbortLineSignal } from '../helpers/abort'
 
-import { parseCronExpression } from "cron-schedule";
-
-/**
- * @description Higher Order BotAction that accepts a ConditionalBotAction (pipeable, that returns a boolean) and based on what boolean it resolves,
- *              does it run the BotAction's provided in a pipe()(). Never the less, it does not return the final ran BotAction return value (if any)
- * @example     givenThat(isGuest)(login(...), closeSomePostLoginModal(),... more BotAction's)
- *              The condition function is async and provided the Puppeteer page instance so you can use it to determine TRUE/FALSE
- * @param condition
- */
-export const givenThat =
-  (condition: ConditionalBotAction) =>
-    (...actions: BotAction<PipeValue|void|AbortLineSignal>[]): BotAction<any> =>
-      async(page, ...injects) => {
-        const resolvedConditionValue: AbortLineSignal|boolean = await condition(page, ...pipeInjects(injects))
-
-        if (isAbortLineSignal(resolvedConditionValue)) {
-          return processAbortLineSignal(resolvedConditionValue)
-        }
-
-        if (resolvedConditionValue) {
-          const returnValue: PipeValue|AbortLineSignal = await pipe()(...actions)(page, ...injects)
-
-          // For consistency, the AbortLineSignal could be processed here, but there is little to gain
-          // from requiring 2 assembledLines to break out of a conditional "if then" block.
-          // Once the then block has been aborted, there is nothing left to run
-          if (isAbortLineSignal(returnValue)) {
-            return returnValue
-          }
-        }
-      }
-
 /**
  * @future support piping in the `collection`
  * @description   A forEach method to loop a collection (array or object's key/value pairs) with assembled BotActions in a pipe
@@ -214,40 +183,3 @@ export const forAsLong =
       }
 
 
-/**
- *
- * @param schedule
- */
-type wildcard = '*'
-type anyValue = wildcard | number | string
-
-// type rangeOfValues = `${number}-${number}`
-// type stepValues = `${rangeOfValues}/${number}` | `${anyValue}/${number}`
-
-type minute = anyValue
-type hour = anyValue
-type dayOfMonth = anyValue
-type month = anyValue
-type weekDay = anyValue
-
-// type minute = anyValue // 0-59
-// type hour = anyValue | stepValues // 0-23
-// type dayOfMonth = anyValue | stepValues // 1-31
-// type month = anyValue | stepValues | 'jan' | 'feb' | 'mar' | 'apr' | 'may' | 'jun' | 'jul' | 'aug' | 'sep' | 'oct' | 'nov' | 'dec' | 'JAN' | 'FEB' | 'MAR' | 'APR' | 'MAY' | 'JUN' | 'JUL' | 'AUG' | 'SEP' | 'OCT' | 'NOV' | 'DEC'
-// type weekDay = anyValue | stepValues | 'mon' | 'tue' | 'wed' | 'thu' | 'fri' | 'sat' | 'sun' | 'MON' | 'TUE' | 'WED' | 'THU' | 'FRI' | 'SAT' | 'SUN'
-
-// much simpler typing, with much faster loading
-type crontab = `${minute} ${hour} ${dayOfMonth} ${month} ${weekDay}`
-
-export const schedule =
-  (schedule: crontab) =>
-    (...actions: BotAction[]): BotAction<any> =>
-      async(page, ...injects) => {
-        // 1. check schedule
-        const cron = parseCronExpression(schedule)
-        console.log(cron.getNextDate(new Date())) // get next date based on now
-
-        // 2. timeout until scheduled run time (with actions set to run)
-
-        // 3. after actions run, reschedule if needed
-      }
